% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htfd.R
\name{htfd_data}
\alias{htfd_data}
\title{Download a Hartford dataset as CSV}
\usage{
htfd_data(key, layer = NULL)
}
\arguments{
\item{key}{Dataset key from htfd_datasets(), or a raw item_id}

\item{layer}{Layer number (only needed if using raw item_id)}
}
\value{
tibble with dataset records
}
\description{
Download a Hartford dataset as CSV
}
