% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htfd.R
\name{htfd_geojson}
\alias{htfd_geojson}
\title{Download Hartford dataset as GeoJSON}
\usage{
htfd_geojson(key, layer = NULL)
}
\arguments{
\item{key}{Dataset key or item_id}

\item{layer}{Layer number (only needed if using raw item_id)}
}
\value{
list (parsed GeoJSON)
}
\description{
Download Hartford dataset as GeoJSON
}
