% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htfd.R
\name{htfd_search}
\alias{htfd_search}
\title{Search Hartford open data portal}
\usage{
htfd_search(q = "")
}
\arguments{
\item{q}{Search query}
}
\value{
tibble: id, type, title, description, url
}
\description{
Search Hartford open data portal
}
